refactor(views): remove age field from member fields partial
The age field has been removed from the visits/_member_fields.html.erb partial. This change was made to simplify the form and potentially streamline data collection where age is no longer required or relevant for the context of this form. Make sure to update any logic that depended on this field elsewhere in the application if necessary.